Nope, still got memory problems and the progress bar doesn't work after the switch...yet.
