+++ app/src/main/java/com/legendbois/memeindexer/ui/main/IndexBuilderFragment.kt
@@ -37,11 +37,20 @@ import kotlin.coroutines.CoroutineContext
 
 class IndexBuilderFragment: Fragment(), View.OnClickListener {
     private var progressNumber: Int = -1
-    private lateinit var model : TextRecognizer
+    internal lateinit var callback: DataProcessor
+
+    fun setDataProcessor(callback: DataProcessor){
+        this.callback = callback
+    }
+
+    interface DataProcessor{
+        fun processData(parentUri: Uri): Boolean
+    }
+    //private lateinit var model : TextRecognizer
     companion object{
         const val TAG="IndexBuilderFragment"
         const val DIRECTORY_REQUEST_CODE=2
-        val imagesRegex="image/.*".toRegex()
+
         fun newInstance(): IndexBuilderFragment{
             return IndexBuilderFragment()
         }
@@ -71,10 +80,10 @@ class IndexBuilderFragment: Fragment(), View.OnClickListener {
                     if (this.context != null) {
                         toggleButtonState(false)
                         val parentUri = data.data
-                        model = TextRecognition.getClient()
+                        //model = TextRecognition.getClient()
                         lifecycleScope.launch {
                             whenStarted {
-                                val complete = processData(parentUri!!)
+                                val complete = callback.processData(parentUri!!)
                             }
                             toggleButtonState(true)
                         }
@@ -85,86 +94,9 @@ class IndexBuilderFragment: Fragment(), View.OnClickListener {
         }
     }
 
-    suspend fun processData(parentUri: Uri): Boolean{
-        withContext(Dispatchers.Default){
-            val db = MemeFilesDatabase.getDatabase(context!!).memeFileDao
-            updateProgressText()
-            traverseDirectoryEntries(parentUri, db)
-            Log.v(TAG, parentUri.toString())
-        }
-        model.close()
-        return true
-    }
 
-    //Thanks to https://stackoverflow.com/questions/41096332/issues-traversing-through-directory-hierarchy-with-android-storage-access-framew
-    fun traverseDirectoryEntries(rootUri: Uri?, db: MemeFileDao){
-        val contentResolver = activity!!.contentResolver
-        var childrenUri: Uri = try {
-            //for childs and sub child dirs
-            DocumentsContract.buildChildDocumentsUriUsingTree(
-                rootUri,
-                DocumentsContract.getDocumentId(rootUri)
-            )
-        } catch (e: java.lang.Exception) {
-            // for parent dir
-            DocumentsContract.buildChildDocumentsUriUsingTree(
-                rootUri,
-                DocumentsContract.getTreeDocumentId(rootUri)
-            )
-        }
 
-        // Keep track of our directory hierarchy
-        val dirNodes: MutableList<Uri> = LinkedList()
-        dirNodes.add(childrenUri)
-        while (dirNodes.isNotEmpty()) {
-            childrenUri = dirNodes.removeAt(0) // get the item from top
-            Log.d(TAG, "node uri:  $childrenUri")
-            val c: Cursor? = contentResolver.query(
-                childrenUri,
-                arrayOf(
-                    DocumentsContract.Document.COLUMN_DOCUMENT_ID,
-                    DocumentsContract.Document.COLUMN_DISPLAY_NAME,
-                    DocumentsContract.Document.COLUMN_MIME_TYPE
-                ),
-                null,
-                null,
-                null
-            )
-            try {
-                if (c!=null) {
-                    while (c.moveToNext()) {
-                        val docId: String = c.getString(0)
-                        val name: String = c.getString(1)
-                        val mime: String = c.getString(2)
-                        if (imagesRegex.matches(mime)){
-                            //Log.v(TAG, "$docId, $name, $mime")
-                            getImageText(DocumentsContract.buildDocumentUriUsingTree(rootUri, docId), name, db)
-                        }
-                        if (isDirectory(mime)) {
-                            val newNode: Uri =
-                                DocumentsContract.buildChildDocumentsUriUsingTree(rootUri, docId)
-                            dirNodes.add(newNode)
-                        }
-                    }
-                }
-            } finally {
-                closeQuietly(c)
-            }
-        }
-        //Room funcs testing
-        /*for (f in db.getAll()){
-            Log.d(TAG, "TestingAll ${f.rowid}, ${f.filename}, ${f.fileuri}")
-        }
 
-        for (f in db.loadTopByFilename("%3000d80%")){
-            Log.d(TAG, "TestingFilename ${f.rowid}, ${f.filename}, ${f.fileuri}")
-        }
-
-        for (f in db.loadTopByText("%maroon 5%")){
-            Log.d(TAG, "TestingText ${f.rowid}, ${f.filename}, ${f.fileuri}")
-        }*/
-
-    }
 
     private fun toggleButtonState(value: Boolean){
         if(value){
@@ -179,80 +111,12 @@ class IndexBuilderFragment: Fragment(), View.OnClickListener {
         indexbuilder_button.isClickable=value
     }
 
-    private fun updateProgressText(){
+    fun updateProgressText(){
         progressNumber+=1
         indexbuilder_progressText?.text="$progressNumber"
     }
 
-    // Util method to check if the mime type is a directory
-    private fun isDirectory(mimeType: String): Boolean {
-        return DocumentsContract.Document.MIME_TYPE_DIR == mimeType
-    }
-    // Util method to close a closeable
-    private fun closeQuietly(closeable: Closeable?) {
-        if (closeable != null) {
-            try {
-                closeable.close()
-            } catch (re: RuntimeException) {
-                throw re
-            } catch (ignore: Exception) {
-                // ignore exception
-            }
-        }
-    }
 
-    private fun getImageText(imageUri: Uri, name: String, db: MemeFileDao){
-        var inpimage : InputImage? = null
-        try {
-            inpimage = InputImage.fromFilePath(activity!!.applicationContext, imageUri)
 
-            model.process(inpimage)
-                .addOnSuccessListener { visionText ->
-                    /*Log.v(
-                        TAG,
-                        "docId: $id, name: $name, text: ${visionText.text}, uri: $imageUri"
-                    )*/
-                    val fileuri = imageUri.toString()
-                    updateProgressText()
-                    if(visionText.text.isNotBlank()) {
-                        val duplicates = db.findUri(fileuri)
-                        if(duplicates.isEmpty()){
-                            db.insert(
-                                MemeFile(
-                                    filename = name,
-                                    fileuri = imageUri.toString(),
-                                    ocrtext = visionText.text.toLowerCase()
-                                )
-                            )
-                        }
-                        else{
-                            for(duplicate in duplicates){
-                                db.update(
-                                    MemeFile(
-                                        rowid = duplicate.rowid,
-                                        fileuri = duplicate.fileuri,
-                                        filename = name,
-                                        ocrtext = visionText.text.toLowerCase()
-                                    )
-                                )
-                            }
-                        }
 
-                    }
-
-                }
-                .addOnFailureListener { e ->
-                    Toast.makeText(activity!!.applicationContext, e.message, Toast.LENGTH_SHORT).show()
-                }
-                .addOnCompleteListener {
-                    inpimage!!.bitmapInternal?.recycle()
-                    inpimage = null
-                }
-        }
-        catch (e: java.lang.Exception){
-            inpimage?.bitmapInternal?.recycle()
-            inpimage = null
-            Log.e(TAG, e.localizedMessage)
-        }
-    }
 }
